ref(issues): Refactor group seen by component
Refactor group seen by component to use group from props instead of
context. Ensures this component will be usable outside of a project
context.
