+++ tests/js/spec/views/groupDetails/seenBy.spec.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import {shallow} from 'enzyme';
 
-import GroupSeenBy from 'app/views/groupDetails/project/seenBy';
+import GroupSeenBy from 'app/views/groupDetails/shared/seenBy';
 import ConfigStore from 'app/stores/configStore';
 
 describe('GroupSeenBy', function() {
@@ -19,29 +19,21 @@ describe('GroupSeenBy', function() {
 
   describe('render()', function() {
     it('should return null if seenBy is falsy', function() {
-      let wrapper = shallow(<GroupSeenBy />, {
-        context: {
-          group: TestStubs.Group({seenBy: undefined}),
-          project: TestStubs.Project(),
-          team: TestStubs.Team(),
-        },
-      });
+      let wrapper = shallow(<GroupSeenBy group={TestStubs.Group({seenBy: undefined})} />);
       expect(wrapper.children()).toHaveLength(0);
     });
 
     it('should return a list of each user that saw', function() {
-      let wrapper = shallow(<GroupSeenBy />, {
-        context: {
-          group: TestStubs.Group({
+      let wrapper = shallow(
+        <GroupSeenBy
+          group={TestStubs.Group({
             seenBy: [
               {id: '1', email: 'jane@example.com'},
               {id: '2', email: 'john@example.com'},
             ],
-          }),
-          project: TestStubs.Project(),
-          team: TestStubs.Team(),
-        },
-      });
+          })}
+        />
+      );
 
       expect(wrapper.find('EyeIcon')).toHaveLength(1);
       expect(wrapper.find('AvatarList')).toHaveLength(1);
